==> The payoff of investing in mutual funds with a high CSR/00-manual.R <==
# 01-update data to 2018 
# 两个data: portfolio.RData and fund-NAV.RData
# 数据来源csmar 2004至2018，更新至2018
# 还有category 

# 02-variables fund characterstic
# data: variables.RData

# 02-variables alpha
# 加入了alpha
# variablesfund.RData

# 02-variables firm characterstic
# data: variablesfirm.RData
# 其中数据已经从2008-2017（由CSR决定），不过CSR为NA的还没有删除
